ajustes finais

+++ src/controllers/Asset.ts
@@ -1,17 +1,20 @@
 import { Request, Response } from "express";
-import mongoose, { Types } from "mongoose";
+import { isValidObjectId } from "mongoose";
 import { Asset, IAsset } from "../database/Asset";
 import { Unit } from "../database/Unit";
 
 const AssetController = {
   async index(req: Request, res: Response): Promise<Response> {
-    let assets = await Asset.find().populate("unit");
+    const assets = await Asset.find().populate("unit");
     return res.status(200).json(assets);
   },
 
   async findById(req: Request, res: Response): Promise<Response> {
     const { id } = req.params;
-    let assets = await Asset.findById(id).populate("unit");
+    if (!isValidObjectId(id)) {
+      return res.status(404).json({ message: "Id is not valid ObjectId" });
+    }
+    const assets = await Asset.findById(id).populate("unit");
     if (!assets) {
       return res.status(404).json({ message: "Asset not found" });
     }
@@ -56,11 +59,23 @@ const AssetController = {
 
   async update(req: Request, res: Response): Promise<Response> {
     const { id } = req.params;
-    const { img, name, description, model, owner, status, healthLevel, unitId } =
-      req.body;
-    
+    if (!isValidObjectId(id)) {
+      return res.status(404).json({ message: "Id is not valid ObjectId" });
+    }
+
+    const {
+      img,
+      name,
+      description,
+      model,
+      owner,
+      status,
+      healthLevel,
+      unitId,
+    } = req.body;
+
     const unit = await Unit.findById(unitId);
-    if(unitId && !unit){
+    if (unitId && !unit) {
       return res.status(404).json({ error: "Unit not found" });
     }
     await Asset.findByIdAndUpdate(id, {
@@ -84,8 +99,12 @@ const AssetController = {
 
   async delete(req: Request, res: Response): Promise<Response> {
     const { id } = req.params;
+
+    if (!isValidObjectId(id)) {
+      return res.status(404).json({ message: "Id is not valid ObjectId" });
+    }
     await Asset.findByIdAndDelete(id)
-      .then((data) => {
+      .then(() => {
         return res.json({ message: `Asset ${id} successfully deleted!` });
       })
       .catch((error) => {
